+++ CHANGELOG.md
@@ -6,17 +6,21 @@
   * CSS has been updated to account for this and no changes should be needed for steps front-end implementation or use as of *0.9.11*
 * Formally introduces the `NfgUi::Components::Foundations::Image` component
   * Images are tooltippable (`tooltip: 'My tooltip'`)
+  * Images accept `:image` option for the path to the image.
+  * Images accept `:src` which is essentially an alias for `:image` (we prefer `:image` since the `Image` component is used throughout the gem and `image: '...'` as an option for other components is more friendly than `src: '...'`)
   * Utilizes a `:responsive` option which is turned on by default.
-    * Turning `:responsive` off (`responsive: false`) removes the `img-fluid` css class.
+    * Turning `:responsive` off (`responsive: false`) removes the `.img-fluid` css class.
 * Introduced the *illustration* component to the `NetworkForGood` system: `NfgUi::Components::Foundations::Illustration` -- this component inherits `NfgUi::Components::Foundations::Image`
   * Illustrations are tooltippable (`tooltip: 'My tooltip'`)
-  * Utilizes sizing options (`size: :lg`) and traits (`ui.nfg :illustration, :lg ...`) as well as the `:image` option.
+  * Utilizes sizing options (`size: :lg`) and traits (`ui.nfg :illustration, :lg ...`)
+  * `:image` option is available for passing in the path to the image. The gem automatically applies the `image_path()` asset pipeline helper method for the image. You're good to go.
   * By default, since sizing is in use, the `:responsive` option is automatically set to `false`, this can be turned on as needed. *Do not pass in a :size option / trait if `responsive: true`*
   * Example usage:
     * `= ui.nfg :illustration, image: 'test_app/sample_illustration.png'`
     * `= ui.nfg :illustration, :lg, image: 'test_app/sample_illustration.png'`
     * `= ui.nfg :illustration, :sm, image: 'test_app/sample_illustration.png'`
-    * `= ui.nfg :illustration, image: 'test_app/sample_illustration.png', responsive: true`
+    * As an option: `= ui.nfg :illustration, image: 'test_app/sample_illustration.png', responsive: true`
+    * As a trait: `= ui.nfg :illustration, :responsive, image: 'test_app/sample_illustration.png'`
 
 # 0.9.10
 *Non-breaking changes*